Remove custom app icons
