+++ package.json
@@ -14,10 +14,8 @@
     "lint": "expo lint"
   },
   "dependencies": {
-    "@config-plugins/react-native-dynamic-app-icon": "^8.0.0",
     "@expo/react-native-action-sheet": "^4.1.1",
     "@expo/vector-icons": "^15.0.2",
-    "@mozzius/expo-dynamic-app-icon": "^1.8.0",
     "@react-native-async-storage/async-storage": "2.2.0",
     "@react-native-community/netinfo": "11.4.1",
     "@react-native-masked-view/masked-view": "0.3.2",
